+++ Sources/SwiftPriorityQueue.swift
@@ -2,7 +2,7 @@
 //  SwiftPriorityQueue.swift
 //  SwiftPriorityQueue
 //
-//  Copyright (c) 2015-2016 David Kopec
+//  Copyright (c) 2015 David Kopec
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -48,7 +48,13 @@ public struct PriorityQueue<T: Comparable> {
             ordered = { $0 < $1 }
         }
         
-        for value in startingValues { push(value) }
+        // Based on "Heap construction" from Sedgewick p 323
+        heap = startingValues
+        var i = heap.count/2 - 1
+        while i >= 0 {
+            sink(i)
+            i -= 1
+        }
     }
     
     /// How many elements the Priority Queue stores
@@ -60,7 +66,7 @@ public struct PriorityQueue<T: Comparable> {
     /// Add a new element onto the Priority Queue. O(lg n)
     ///
     /// - parameter element: The element to be inserted into the Priority Queue.
-    public mutating func push(element: T) {
+    public mutating func push(_ element: T) {
         heap.append(element)
         swim(heap.count - 1)
     }
@@ -97,7 +103,7 @@ public struct PriorityQueue<T: Comparable> {
     }
     
     // Based on example from Sedgewick p 316
-    private mutating func sink(index: Int) {
+    private mutating func sink(_ index: Int) {
         var index = index
         while 2 * index + 1 < heap.count {
             
@@ -112,7 +118,7 @@ public struct PriorityQueue<T: Comparable> {
     }
     
     // Based on example from Sedgewick p 316
-    private mutating func swim(index: Int) {
+    private mutating func swim(_ index: Int) {
         var index = index
         while index > 0 && ordered(heap[(index - 1) / 2], heap[index]) {
             swap(&heap[(index - 1) / 2], &heap[index])
@@ -131,8 +137,8 @@ extension PriorityQueue: IteratorProtocol {
 // MARK: - SequenceType
 extension PriorityQueue: Sequence {
     
-    public typealias Generator = PriorityQueue
-    public func generate() -> Generator { return self }
+    public typealias Iterator = PriorityQueue
+    public func makeIterator() -> Iterator { return self }
 }
 
 // MARK: - CollectionType
@@ -144,6 +150,12 @@ extension PriorityQueue: Collection {
     public var endIndex: Int { return heap.endIndex }
     
     public subscript(i: Int) -> T { return heap[i] }
+    
+    #if swift(>=3.0)
+        public func index(after i: PriorityQueue.Index) -> PriorityQueue.Index {
+            return heap.index(after: i)
+        }
+    #endif
 }
 
 // MARK: - CustomStringConvertible, CustomDebugStringConvertible